requirements.txt
# Core dependencies
agno>=0.15.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0.3
httpx>=0.24.1

# Memory and vector search
faiss-cpu>=1.7.4
lancedb>=0.3.0
pydantic>=2.0.0

# Monitoring and metrics
prometheus-client>=0.17.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0